requirements.txt
gymnasium==1.1.1
matplotlib==3.9.4
networkx==3.4.2
numpy==2.2.4
PyYAML==6.0.2
stable_baselines3==2.6.0
torch==2.6.0
xmltodict==0.14.2

skrl~=1.4.3